implementação da listagem das minhas viagens Refatora activitys de home e mytrips, movendo para suas presenters
