--- app/src/main/java/com/example/trabalho/MyTripsActivityTest.java
@@ -1,14 +1,0 @@
-package com.example.trabalho;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MyTripsActivityTest extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my_trips_test);
-    }
-}
