+++ app/src/main/java/com/example/trabalho/ProfileActivity.java
@@ -2,33 +2,13 @@ package com.example.trabalho;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
-import android.widget.Toast;
 
-import com.example.trabalho.presenter.contracts.ActivityContract;
-
-public class ProfileActivity extends AppCompatActivity implements ActivityContract.ActivityView {
+public class ProfileActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_profile);
-    }
-
-    @Override
-    public Context getContext() {
-        return this.getApplicationContext();
-    }
-
-    @Override
-    public void showToast(String message) {
-        Toast.makeText(this.getApplicationContext(), message, Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void navigate(Intent intent) {
-        startActivity(intent);
     }
 }
