--- app/src/main/res/layout/show_my_trips.xml
@@ -1,135 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android">
-    <data>
-        <import type="com.example.trabalho.utils.Helper"/>
-        <variable
-            name="tripModel"
-            type="com.example.trabalho.models.Trip" />
-        <variable
-            name="myTripsAdapter"
-            type="com.example.trabalho.adapters.MyTripsAdapter" />
-    </data>
-
-    <LinearLayout
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:padding="10dp">
-
-        <androidx.cardview.widget.CardView
-            android:id="@+id/cardTrip"
-            android:layout_width="388dp"
-            android:layout_height="234dp"
-            android:tag="@{tripModel}"
-            app:contentPadding="5dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <LinearLayout
-                android:layout_width="382dp"
-                android:layout_height="234dp"
-                android:orientation="vertical">
-
-                <TextView
-                    android:id="@+id/textView20"
-                    android:layout_width="112dp"
-                    android:layout_height="21dp"
-                    android:layout_marginTop="16dp"
-                    android:text="@{tripModel.place}"
-                    android:textColor="#FFFFFF"
-                    android:textSize="14sp"
-                    app:layout_constraintEnd_toEndOf="@+id/view"
-                    app:layout_constraintHorizontal_bias="0.927"
-                    app:layout_constraintStart_toStartOf="@+id/view"
-                    app:layout_constraintTop_toTopOf="@+id/view" />
-
-                <TextView
-                    android:id="@+id/textView19"
-                    android:layout_width="102dp"
-                    android:layout_height="19dp"
-                    android:layout_marginTop="32dp"
-                    android:text="@{Helper.getFormattedDate(tripModel.departureDate)}"
-                    android:textColor="#FFFFFF"
-                    android:textSize="14sp"
-                    app:layout_constraintEnd_toStartOf="@+id/imageView3"
-                    app:layout_constraintHorizontal_bias="0.756"
-                    app:layout_constraintStart_toStartOf="@+id/view"
-                    app:layout_constraintTop_toBottomOf="@+id/textView10" />
-
-                <TextView
-                    android:id="@+id/textView21"
-                    android:layout_width="112dp"
-                    android:layout_height="21dp"
-                    android:layout_marginTop="12dp"
-                    android:text="@{Helper.getFormattedDate(tripModel.arrivalDate)}"
-                    android:textColor="#FFFFFF"
-                    android:textSize="14sp"
-                    app:layout_constraintEnd_toEndOf="@+id/view"
-                    app:layout_constraintHorizontal_bias="0.527"
-                    app:layout_constraintStart_toEndOf="@+id/imageView3"
-                    app:layout_constraintTop_toBottomOf="@+id/textView20" />
-
-                <TextView
-                    android:id="@+id/textView22"
-                    android:layout_width="112dp"
-                    android:layout_height="21dp"
-                    android:layout_marginTop="12dp"
-                    android:text="@{Helper.getFormattedDate(tripModel.returnDate)}"
-                    android:textColor="#FFFFFF"
-                    android:textSize="14sp"
-                    app:layout_constraintEnd_toEndOf="@+id/view"
-                    app:layout_constraintHorizontal_bias="0.527"
-                    app:layout_constraintStart_toEndOf="@+id/imageView3"
-                    app:layout_constraintTop_toBottomOf="@+id/textView21" />
-
-                <ImageView
-                    android:id="@+id/imageView3"
-                    android:layout_width="32dp"
-                    android:layout_height="32dp"
-                    android:layout_marginTop="28dp"
-                    android:foregroundGravity="center"
-                    app:layout_constraintEnd_toStartOf="@+id/textView20"
-                    app:layout_constraintHorizontal_bias="0.851"
-                    app:layout_constraintStart_toStartOf="@+id/view"
-                    app:layout_constraintTop_toTopOf="parent"
-                    app:srcCompat="@drawable/airplane" />
-
-                <Button
-                    android:id="@+id/add_new_trip_button"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:background="@drawable/button_my_trips"
-                    android:gravity="center"
-                    android:padding="8dp"
-                    android:text="@string/see_details"
-                    android:textColor="#FFFFFF"
-                    app:backgroundTint="#4A90E2"
-                    app:layout_constraintBottom_toBottomOf="@+id/view"
-                    app:layout_constraintEnd_toEndOf="@+id/view"
-                    app:layout_constraintStart_toStartOf="@+id/view"
-                    app:layout_constraintTop_toBottomOf="@+id/imageView3"
-                    app:layout_constraintVertical_bias="1.0" />
-
-                <TextView
-                    android:id="@+id/textView10"
-                    android:layout_width="104dp"
-                    android:layout_height="21dp"
-                    android:layout_marginTop="16dp"
-                    android:text="@{tripModel.place}"
-                    android:textColor="#FFFFFF"
-                    android:textSize="14sp"
-                    app:layout_constraintEnd_toStartOf="@+id/imageView3"
-                    app:layout_constraintHorizontal_bias="0.794"
-                    app:layout_constraintStart_toStartOf="@+id/view"
-                    app:layout_constraintTop_toTopOf="@+id/view" />
-
-            </LinearLayout>
-
-        </androidx.cardview.widget.CardView>
-
-    </LinearLayout>
-</layout>
